feat: actualizar servicio de roles para mostrar un listado de roles

+++ src/modules/admin/rol/rol.service.ts
@@ -11,7 +11,7 @@ import { PrismaService } from 'src/prisma/prisma.service';
 import { handleException } from 'src/utils';
 import { ValidRols } from '../auth/interfaces';
 import { UpdateRolDto } from './dto/update-rol.dto';
-import { HttpResponse, RolPermissions, RolModulesPermissions, Rol } from 'src/interfaces';
+import { HttpResponse, RolPermissions, RolModulesPermissions, Rol, UserData } from 'src/interfaces';
 
 @Injectable()
 export class RolService {
@@ -26,7 +26,7 @@ export class RolService {
    */
   async create(createRolDto: CreateRolDto): Promise<HttpResponse<Rol>> {
     try {
-      const { name, description, modulePermissions } = createRolDto;
+      const { name, description, rolPermissions } = createRolDto;
 
       // Verificar si el rol ya existe y está activo
       const rolExist = await this.prisma.rol.findUnique({
@@ -54,8 +54,8 @@ export class RolService {
         });
 
         // Verificar existencia de permisos y construir las entradas para RolModulePermissions
-        const rolModulePermissionEntries = [];
-        for (const modulePermissionId of modulePermissions) {
+        const rolPermissionEntries = [];
+        for (const modulePermissionId of rolPermissions) {
           const modulePermission = await prisma.modulePermissions.findUnique({
             where: {
               id: modulePermissionId
@@ -68,7 +68,7 @@ export class RolService {
             );
           }
 
-          rolModulePermissionEntries.push({
+          rolPermissionEntries.push({
             rolId: createdRol.id,
             modulePermissionsId: modulePermissionId
           });
@@ -76,7 +76,7 @@ export class RolService {
 
         // Crear las relaciones entre rol y permisos en una transacción
         await prisma.rolModulePermissions.createMany({
-          data: rolModulePermissionEntries,
+          data: rolPermissionEntries,
           skipDuplicates: true
         });
 
@@ -107,7 +107,7 @@ export class RolService {
    */
   async update(id: string, updateRolDto: UpdateRolDto): Promise<HttpResponse<RolPermissions>> {
     try {
-      const { name, description, modulePermissions } = updateRolDto;
+      const { name, description, rolPermissions } = updateRolDto;
 
       // Validar que el rol no sea SUPER_ADMIN
       const rolIsSuperAdmin = await this.isRolSuperAdmin(id);
@@ -164,8 +164,8 @@ export class RolService {
 
         // Validar existencia de permisos y construir nuevas entradas
         const rolModulePermissionEntries = [];
-        if (modulePermissions && modulePermissions.length > 0) {
-          for (const modulePermissionId of modulePermissions) {
+        if (rolPermissions && rolPermissions.length > 0) {
+          for (const modulePermissionId of rolPermissions) {
             const modulePermission = await prisma.modulePermissions.findUnique({
               where: {
                 id: modulePermissionId
@@ -210,7 +210,7 @@ export class RolService {
         }
 
         // Devolver datos actualizados del rol
-        const rolPermissions = await prisma.rolModulePermissions.findMany({
+        const rolPermissionsDB = await prisma.rolModulePermissions.findMany({
           where: { rolId: id },
           select: {
             modulePermissions: {
@@ -237,7 +237,7 @@ export class RolService {
         });
 
         // Agrupar permisos por módulo
-        const groupedByModule = rolPermissions.reduce((acc, rolModulePermission) => {
+        const groupedByModule = rolPermissionsDB.reduce((acc, rolModulePermission) => {
           const moduleId = rolModulePermission.modulePermissions.module.id;
           const module = rolModulePermission.modulePermissions.module;
           const permission = rolModulePermission.modulePermissions.permission;
@@ -283,7 +283,7 @@ export class RolService {
    * @param id Id del rol a eliminar
    * @returns  Datos del rol eliminado
    */
-  async remove(id: string): Promise<HttpResponse<RolPermissions>> {
+  async remove(id: string): Promise<HttpResponse<Rol>> {
     try {
       // Validar fuera de la transacción
       const rolIsSuperAdmin = await this.isRolSuperAdmin(id);
@@ -298,57 +298,29 @@ export class RolService {
         throw new BadRequestException('It is not possible to delete the rol because it is in use');
       }
 
+      const rolIsUsedByInactiveUsers = await this.rolIsUsedByInactiveUsers(id);
+
       // Inicia la transacción
       const removedRol = await this.prisma.$transaction(async (prisma) => {
-        // Eliminar el rol, y las relaciones en RolModulePermissions se eliminarán automáticamente por la cascada
-        const rolDelete = await prisma.rol.delete({
-          where: { id },
-          select: {
-            id: true,
-            name: true,
-            description: true,
-            rolModulePermissions: {
-              select: {
-                id: true,
-                modulePermissions: {
-                  select: {
-                    id: true,
-                    module: {
-                      select: {
-                        id: true,
-                        cod: true,
-                        name: true,
-                        description: true
-                      }
-                    },
-                    permission: {
-                      select: {
-                        id: true,
-                        cod: true,
-                        name: true,
-                        description: true
-                      }
-                    }
-                  }
-                }
-              }
+        //Eliminar el rol, y las relaciones en RolModulePermissions se eliminarán automáticamente por la cascada
+        let rolDelete: Rol;
+        if (rolIsUsedByInactiveUsers) {
+          rolDelete = await prisma.rol.update({
+            where: { id },
+            data: {
+              isActive: false
             }
-          }
-        });
+          });
+        } else {
+          rolDelete = await prisma.rol.delete({
+            where: { id }
+          });
+        }
 
         return {
           id: rolDelete.id,
           name: rolDelete.name,
-          description: rolDelete.description,
-          rolPermissions: rolDelete.rolModulePermissions.map((rolModulePermission) => ({
-            module: {
-              id: rolModulePermission.modulePermissions.module.id,
-              cod: rolModulePermission.modulePermissions.module.cod,
-              name: rolModulePermission.modulePermissions.module.name,
-              description: rolModulePermission.modulePermissions.module.description
-            },
-            permissions: [rolModulePermission.modulePermissions.permission]
-          }))
+          description: rolDelete.description
         };
       });
 
@@ -370,12 +342,12 @@ export class RolService {
    * Obtener todos los roles con sus módulos y permisos
    * @returns Lista de roles con módulos y permisos
    */
-  async findAll(): Promise<any[]> {
+  async findAll(user: UserData): Promise<RolPermissions[]> {
     try {
       // Recupera todos los roles activos con sus permisos asociados
       const roles = await this.prisma.rol.findMany({
         where: {
-          isActive: true,
+          ...(user.isSuperAdmin ? {} : { isActive: true }),
           NOT: {
             name: ValidRols.SUPER_ADMIN
           }
@@ -391,6 +363,9 @@ export class RolService {
               }
             }
           }
+        },
+        orderBy: {
+          createdAt: 'asc'
         }
       });
 
@@ -413,15 +388,17 @@ export class RolService {
             });
           }
 
-          modulesMap
-            .get(moduleId)
-            ?.permissions.push(rolModulePermission.modulePermissions.permission);
+          modulesMap.get(moduleId)?.permissions.push({
+            ...rolModulePermission.modulePermissions.permission,
+            idModulePermission: rolModulePermission.modulePermissions.id
+          });
         });
 
         return {
           id: role.id,
           name: role.name,
           description: role.description,
+          isActive: role.isActive,
           rolPermissions: Array.from(modulesMap.values())
         };
       });
@@ -508,13 +485,50 @@ export class RolService {
   }
 
   /**
+   * Mostrar todos los módulos con sus permisos
+   * @returns Lista de módulos con sus permisos
+   */
+  async findAllModulesPermissions(): Promise<RolModulesPermissions[]> {
+    try {
+      // Obtener todos los módulos con sus permisos
+      const modulesPermissions = await this.prisma.modulePermissions.findMany({
+        include: {
+          module: {
+            select: {
+              id: true,
+              cod: true,
+              name: true,
+              description: true
+            }
+          },
+          permission: {
+            select: {
+              id: true,
+              cod: true,
+              name: true,
+              description: true
+            }
+          }
+        },
+        orderBy: {
+          moduleId: 'asc'
+        }
+      });
+      return this.groupPermissionsByModule(modulesPermissions);
+    } catch (error) {
+      this.logger.error('Error getting modules with permissions', error.stack);
+      throw new Error('Error getting modules with permissions');
+    }
+  }
+
+  /**
    * Verificar si no hay datos para actualizar
    * @param updateRolDto Datos del rol a actualizar
    * @returns  True si no hay datos para actualizar, false en caso contrario
    */
   private isNoDataUpdate(updateRolDto: UpdateRolDto): boolean {
     // Implementa la lógica para verificar si no hay datos para actualizar
-    return !updateRolDto.name && !updateRolDto.description && !updateRolDto.modulePermissions;
+    return !updateRolDto.name && !updateRolDto.description && !updateRolDto.rolPermissions;
   }
 
   /**
@@ -565,10 +579,69 @@ export class RolService {
     const rolIsUsed = await this.prisma.userRol.findFirst({
       where: {
         rolId: id,
-        isActive: true
+        isActive: true,
+        user: {
+          isActive: true
+        }
       }
     });
 
     return !!rolIsUsed;
+  }
+
+  /**
+   * Verificar si el rol esta en uso, pero por usuarios inactivos
+   * @param id Id del rol a verificar si esta en uso
+   * @returns True si el rol esta en uso por usuarios inactivos, false en caso contrario
+   */
+  async rolIsUsedByInactiveUsers(id: string): Promise<boolean> {
+    const rolIsUsed = await this.prisma.user.findMany({
+      where: {
+        isActive: false,
+        userRols: {
+          some: {
+            rolId: id
+          }
+        }
+      }
+    });
+    // Devuelve true si el rol está en uso por usuarios inactivos
+    return rolIsUsed.length > 0;
+  }
+
+  /**
+   * Agrupar permisos por modulos
+   * @param rolModulePermissions Permisos de un rol
+   * @returns Permisos agrupados por módulos
+   */
+  private groupPermissionsByModule(rolModulePermissions: any[]): RolModulesPermissions[] {
+    return rolModulePermissions.reduce((acc, entry) => {
+      const { moduleId, module, permission, id } = entry;
+      const modulePermission = {
+        idModulePermission: id,
+        ...permission
+      };
+
+      // Verificar si ya existe el módulo en el resultado
+      const existingModule = acc.find((item) => item.module.id === moduleId);
+
+      if (existingModule) {
+        // Si ya existe, añadir el permiso al array de permisos
+        existingModule.permissions.push(modulePermission);
+      } else {
+        // Si no existe, crear un nuevo módulo con el array de permisos
+        acc.push({
+          module: {
+            id: module.id,
+            cod: module.cod,
+            name: module.name,
+            description: module.description
+          },
+          permissions: [modulePermission]
+        });
+      }
+
+      return acc;
+    }, []);
   }
 }
